+++ frontend/src/components/Snapshot/SnapshotSimpleList/tabs/SnapshotAssetGroupPerformanceTab.tsx
@@ -446,7 +446,7 @@ const SnapshotAssetGroupPerformanceTab: React.FC<SnapshotAssetGroupPerformanceTa
           hasPrev={assetGroupPerformanceData.hasPrev}
           onPageChange={onPageChange}
           onLimitChange={onLimitChange}
-          limitOptions={[5, 10, 20, 50]}
+          limitOptions={[5, 10, 25, 50, 100]}
         />
       )}
     </Box>